bukkit-4.13.8 | REALLY! added PacketUtils#getRespawnPacket 1.19 compatibility

+++ bukkit/src/main/java/me/mohamad82/ruom/utils/PacketUtils.java
@@ -44,7 +44,7 @@ public class PacketUtils {
                         nmsOldGameMode,
                         false,
                         isFlat,
-                        0,
+                        (byte) 0,
                         Optional.empty()
                 );
             } else {
